fix(analytics-js): add global definitions extended window type (#1445)
