--- packages/analytics-js-plugins/src/types/rudderanalytics.d.ts
@@ -1,14 +1,0 @@
-import type {
-  IRudderStackGlobals,
-  RudderAnalytics,
-  PreloadedEventCall,
-} from '@rudderstack/analytics-js';
-
-declare global {
-  interface Window {
-    rudderanalytics: RudderAnalytics | PreloadedEventCall[];
-    RudderStackGlobals: IRudderStackGlobals;
-    rudderAnalyticsMount: () => void;
-    rudderAnalyticsBuildType: 'legacy' | 'modern';
-  }
-}
